...G...PS. [ZBX-11122] fixed comparisons between signed and unsigned integer expressions to eliminate build warnings
git-svn-id: svn://svn.zabbix.com/trunk@63171 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/libs/zbxjson/json.c
@@ -653,7 +653,7 @@ static const char	*zbx_json_copy_string(const char *p, char *out, size_t size)
 				*out++ = *p++;
 		}
 
-		if (out - start == size)
+		if ((size_t)(out - start) == size)
 			break;
 	}
